Adds Console clear functions
Adds Console Clear functions
Fixes Layout
Fixes small sorting bug
