+++ Festivity/FestivalRegister.cs
@@ -1,7 +1,6 @@
 ﻿using Newtonsoft.Json;
 using System;
 using System.IO;
-using System.Security.Cryptography.X509Certificates;
 
 namespace Festivity
 {
