+++ Festivity/FestivalsDatabase.json
@@ -2,109 +2,129 @@
   "festivals": [
     {
       "festivalId": 1,
-      "festivalName": "Robin",
+      "festivalName": "Robin's Festival",
       "festivalDate": {
-        "day": 18,
-        "month": 5,
+        "day": 29,
+        "month": 03,
         "year": 2020
       },
       "festivalStartingTime": "1500",
       "festivalEndTime": "1600",
       "festivalLocation": {
         "country": "Nederland",
-        "city": "rotje",
+        "city": "Rotterdam",
         "zipCode": "3066GK",
         "streetName": "123adresstraat",
         "streetNumber": "25"
       },
-      "festivalDescription": "Dit feestje is echt zo bizar cool daar zijn geen woorden voor",
+      "festivalDescription": "This party is hella cool for real",
       "festivalAgeRestriction": 23,
       "festivalGenre": "Drum & Bass"
     },
     {
-      "festivalId": 1,
-      "festivalName": "Dikke fissa maat",
+      "festivalId": 2,
+      "festivalName": "Mark's Festival",
       "festivalDate": {
-        "day": 18,
-        "month": 5,
+        "day": 26,
+        "month": 05,
         "year": 2020
       },
       "festivalStartingTime": "1700",
       "festivalEndTime": "1800",
       "festivalLocation": {
         "country": "Nederland",
-        "city": "rotje",
+        "city": "Rotterdam",
         "zipCode": "3066GK",
         "streetName": "123adresstraat",
         "streetNumber": "25"
       },
-      "festivalDescription": "Dit feestje is echt zo bizar cool daar zijn geen woorden voor",
+      "festivalDescription": "Congratulations Mark",
       "festivalAgeRestriction": 23,
       "festivalGenre": "Drum & Bass"
     },
     {
-      "festivalId": 1,
-      "festivalName": "abcdefghijklmnopqrstuvwxyz",
+      "festivalId": 3,
+      "festivalName": "Enny's Festival",
       "festivalDate": {
-        "day": 18,
-        "month": 5,
+        "day": 1,
+        "month": 1,
+        "year": 2021
+      },
+      "festivalStartingTime": "1999",
+      "festivalEndTime": "2359",
+      "festivalLocation": {
+        "country": "Nederland",
+        "city": "Rotterdam",
+        "zipCode": "3066GK",
+        "streetName": "123adresstraat",
+        "streetNumber": "25"
+      },
+      "festivalDescription": "Cool party with loads of nachos and stuff",
+      "festivalAgeRestriction": 23,
+      "festivalGenre": "Drum & Bass"
+    },
+    {
+      "festivalId": 4,
+      "festivalName": "Kelvin's Festival",
+      "festivalDate": {
+        "day": 12,
+        "month": 10,
         "year": 2020
       },
       "festivalStartingTime": "1999",
       "festivalEndTime": "2359",
       "festivalLocation": {
         "country": "Nederland",
-        "city": "rotje",
+        "city": "Rotterdam",
         "zipCode": "3066GK",
         "streetName": "123adresstraat",
         "streetNumber": "25"
       },
-      "festivalDescription": "Dit feestje is echt zo bizar cool daar zijn geen woorden voor",
+      "festivalDescription": "Kelvin's Festival",
       "festivalAgeRestriction": 23,
       "festivalGenre": "Drum & Bass"
     },
     {
-      "festivalId": 1,
-      "festivalName": "test4",
+      "festivalId": 5,
+      "festivalName": "Samuel's Party",
       "festivalDate": {
-        "day": 19,
-        "month": 5,
+        "day": 30,
+        "month": 06,
         "year": 2020
       },
       "festivalStartingTime": "1999",
       "festivalEndTime": "2359",
       "festivalLocation": {
         "country": "Nederland",
-        "city": "rotje",
+        "city": "Rotterdam",
         "zipCode": "3066GK",
         "streetName": "123adresstraat",
         "streetNumber": "25"
       },
-      "festivalDescription": "Dit feestje is echt zo bizar cool daar zijn geen woorden voor",
+      "festivalDescription": "Samuel's party",
       "festivalAgeRestriction": 23,
       "festivalGenre": "Drum & Bass"
     },
     {
-      "festivalId": 1,
-      "festivalName": "test5",
+      "festivalId": 6,
+      "festivalName": "Page 2 testparty",
       "festivalDate": {
-        "day": 17,
-        "month": 5,
-        "year": 2020
+        "day": 30,
+        "month": 06,
+        "year": 2025
       },
       "festivalStartingTime": "1999",
       "festivalEndTime": "2359",
       "festivalLocation": {
         "country": "Nederland",
-        "city": "rotje",
+        "city": "Rotterdam",
         "zipCode": "3066GK",
         "streetName": "123adresstraat",
         "streetNumber": "25"
       },
-      "festivalDescription": "Dit feestje is echt zo bizar cool daar zijn geen woorden voor",
+      "festivalDescription": "Test description",
       "festivalAgeRestriction": 23,
-      "festivalGenre": "Drum & Bass",
-      "festivalOrganiserId": 1
-    }
+      "festivalGenre": "Drum & Bass"
+    },
   ]
 }
